requirements.txt
fastapi~=0.111.0
python-dotenv~=1.0.1

aiomysql~=0.2.0
aiohttp~=3.9.5
requests~=2.31.0
steam~=1.4.4
tqdm~=4.66.4